Code from main.js divided into separate modules main.js connected as module file
