+++ js/data.js
@@ -1,0 +1,28 @@
+const NAMES = [
+  'Ли',
+  'Клементина',
+  'Кенни',
+  'Криста',
+  'Омид'
+];
+
+const DESCRIPTIONS = [
+  'Здесь должна быть подпись.',
+  'Это моя лучшая фотография, как считаете?',
+  'С новой камерой снимать стало гораздо проще!',
+  'Новый день - новый пост.',
+  'Подпись к фотографии - это целое искусство.'
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const usedID = [];
+
+export {NAMES, DESCRIPTIONS, MESSAGES, usedID};
